Move signalR service under services folder
